Changed up the enrollment scroll service to change the hash after a scroll has happened.
The scroll event will actually be called twice because of the initial scroll and the locationChange event. The locationChange event is set to a animated speed of 0 so that it will just jump to the location. This will handle the browser back/forward buttons and coming into a specific section on page load.